requirements.txt
fastapi
uvicorn[standard]
requests
llama-index-core
llama-index-vector-stores-chroma
llama-index-llms-deepseek
chromadb
supabase
